src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
asgiref==3.8.1
    # via django
boto3==1.34.145
    # via -r src/requirements/requirements.in
botocore==1.34.145
    # via
    #   boto3
    #   s3transfer
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
dj-database-url==2.2.0
    # via -r src/requirements/requirements.in
django==4.1.13
    # via
    #   -r src/requirements/requirements.in
    #   dj-database-url
    #   django-storages
django-storages==1.14.4
    # via -r src/requirements/requirements.in
gunicorn==22.0.0
    # via -r src/requirements/requirements.in
idna==3.7
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
packaging==24.1
    # via gunicorn
pillow==10.4.0
    # via -r src/requirements/requirements.in
psycopg2==2.9.9
    # via -r src/requirements/requirements.in
python-dateutil==2.9.0.post0
    # via botocore
python-decouple==3.8
    # via -r src/requirements/requirements.in
requests==2.32.3
    # via stripe
s3transfer==0.10.2
    # via boto3
six==1.16.0
    # via python-dateutil
sqlparse==0.5.1
    # via django
stripe==10.4.0
    # via -r src/requirements/requirements.in
typing-extensions==4.12.2
    # via
    #   dj-database-url
    #   stripe
tzdata==2024.1
    # via django
urllib3==2.2.2
    # via
    #   botocore
    #   requests
